Change exp type

+++ include/boost/charconv/detail/ryu/ryu_generic_128.hpp
@@ -427,7 +427,7 @@ static inline int generic_to_chars(const struct floating_decimal_128 v, char* re
     // See: https://github.com/cppalliance/charconv/issues/64
     if (fmt == chars_format::general)
     {
-        const int32_t exp = v.exponent + (int32_t)olength;
+        const int64_t exp = v.exponent + (int64_t)olength;
         if (exp <= 0 && exp >= -4)
         {
             return generic_to_chars_fixed(v, result, result_size, precision);
